+++ CHANGELOG.md
@@ -14,8 +14,9 @@ This project adheres to [Semantic Versioning](http://semver.org). See [Keep A Ch
 
 ### Changed
 - Use `ircmaxell/random-lib` to generate random numbers for mobile ID authentication queries
+- Public interface of `Authenticator` has changed: session code is now held internally and is not passed as a parameter to calls after `authenticate`
 
 ## 0.1.0 - 2015-11-04
 
 Initial release that contains prototype mobile ID authentication and file signing functionality through the
-`\Bigbank\DigiDoc\Services\MobileId\Authenticator` and `\Bigbank\DigiDoc\Services\MobileId\FileSigner` services.
+`\Bigbank\DigiDoc\Services\MobileId\Authenticator` and `\Bigbank\DigiDoc\Services\MobileId\FileSigner` services